+++ vectorforge/engine/potrace_engine.py
@@ -156,7 +156,8 @@ def potrace_binary_to_svg(
         svg = (
             f'<?xml version="1.0" encoding="UTF-8"?>\n'
             f'<svg xmlns="http://www.w3.org/2000/svg" '
-            f'width="{w}" height="{h}" viewBox="0 0 {w} {h}">\n'</svg>\n'
+            f'width="{w}" height="{h}" viewBox="0 0 {w} {h}">\n'
+            f'</svg>\n'
         )
         return svg, PathStats(0, 0)